+++ R/backend-redshift.R
@@ -14,6 +14,10 @@ sql_translate_env.RedshiftConnection <- function(con) {
       paste0 = sql_paste_redshift(""),
       str_c = sql_paste_redshift(""),
 
+      # https://docs.aws.amazon.com/redshift/latest/dg/r_SUBSTRING.html
+      substr = sql_substr("SUBSTRING"),
+      str_sub = sql_str_sub("SUBSTRING", "LEN"),
+
       # https://docs.aws.amazon.com/redshift/latest/dg/REGEXP_REPLACE.html
       str_replace = sql_not_supported("str_replace"),
       str_replace_all = function(string, pattern, replacement) {